Improved style.css
Added classes for stuff like dynamic width, a header banner,
navigation menu, etc. that all the big boy websites use.

A lot of it is directly taken from the default CSS of Jekyll. I'm
gonna definitely change it as I start learning more about design and
shit but I needed a decent foundation to work off of and figured this
would be the best course of action for now.