Attempting to fix crashing error
Sometimes the server just crashes seemingly randomly because the
function process_request tries to use split on a nil object. I can't
reproduce this behavior but it seems to be caused by socket.gets
returning a nil value, so I've implemented a check for that within
the function and a way of handling it within servertest. I won't
know for sure if it's worked or not until the crash happens again or
if "nil request" is printed to the console during a test.